requirements.txt
boto3~=1.34.145
pymongo~=4.8.0
python-dotenv~=1.0.1
Flask~=3.0.3
werkzeug~=3.0.3
flask-mail~=0.10.0
